+++ packages/mixin_logger/macos/Classes/mixin_logger.h
@@ -1,1 +1,0 @@
-../../src/include/mixin_logger/mixin_logger.h